requirements.txt
praw>=7.7.1
spotipy>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0